Release 1.10

+++ flink-table/flink-table-api-java/src/main/java/org/apache/flink/table/catalog/CatalogManager.java
@@ -31,11 +31,15 @@ import org.apache.flink.util.StringUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.Arrays;
+import java.util.Collections;
+import java.util.HashMap;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Optional;
 import java.util.Set;
+import java.util.function.Predicate;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 import static java.lang.String.format;
 import static org.apache.flink.util.Preconditions.checkArgument;
@@ -51,6 +55,10 @@ public class CatalogManager {
 
 	// A map between names and catalogs.
 	private Map<String, Catalog> catalogs;
+
+	// Those tables take precedence over corresponding permanent tables, thus they shadow
+	// tables coming from catalogs.
+	private Map<ObjectIdentifier, CatalogBaseTable> temporaryTables;
 
 	// The name of the current catalog and database
 	private String currentCatalogName;
@@ -70,6 +78,7 @@ public class CatalogManager {
 		this.currentCatalogName = defaultCatalogName;
 		this.currentDatabaseName = defaultCatalog.getDefaultDatabase();
 
+		this.temporaryTables = new HashMap<>();
 		// right now the default catalog is always the built-in one
 		this.builtInCatalogName = defaultCatalogName;
 	}
@@ -104,19 +113,10 @@ public class CatalogManager {
 	}
 
 	/**
-	 * Retrieves names of all registered catalogs.
-	 *
-	 * @return a set of names of registered catalogs
-	 */
-	public Set<String> getCatalogs() {
-		return catalogs.keySet();
-	}
-
-	/**
 	 * Gets the current catalog that will be used when resolving table path.
 	 *
 	 * @return the current catalog
-	 * @see CatalogManager#qualifyIdentifier(String...)
+	 * @see CatalogManager#qualifyIdentifier(UnresolvedIdentifier)
 	 */
 	public String getCurrentCatalog() {
 		return currentCatalogName;
@@ -127,7 +127,7 @@ public class CatalogManager {
 	 *
 	 * @param catalogName catalog name to set as current catalog
 	 * @throws CatalogNotExistException thrown if the catalog doesn't exist
-	 * @see CatalogManager#qualifyIdentifier(String...)
+	 * @see CatalogManager#qualifyIdentifier(UnresolvedIdentifier)
 	 */
 	public void setCurrentCatalog(String catalogName) throws CatalogNotExistException {
 		checkArgument(!StringUtils.isNullOrWhitespaceOnly(catalogName), "Catalog name cannot be null or empty.");
@@ -152,7 +152,7 @@ public class CatalogManager {
 	 * Gets the current database name that will be used when resolving table path.
 	 *
 	 * @return the current database
-	 * @see CatalogManager#qualifyIdentifier(String...)
+	 * @see CatalogManager#qualifyIdentifier(UnresolvedIdentifier)
 	 */
 	public String getCurrentDatabase() {
 		return currentDatabaseName;
@@ -164,7 +164,7 @@ public class CatalogManager {
 	 *
 	 * @param databaseName database name to set as current database name
 	 * @throws CatalogException thrown if the database doesn't exist in the current catalog
-	 * @see CatalogManager#qualifyIdentifier(String...)
+	 * @see CatalogManager#qualifyIdentifier(UnresolvedIdentifier)
 	 * @see CatalogManager#setCurrentCatalog(String)
 	 */
 	public void setCurrentDatabase(String databaseName) {
@@ -209,65 +209,235 @@ public class CatalogManager {
 	}
 
 	/**
+	 * Result of a lookup for a table through {@link #getTable(ObjectIdentifier)}. It combines the
+	 * {@link CatalogBaseTable} with additional information such as if the table is a temporary table or comes
+	 * from the catalog.
+	 */
+	public static class TableLookupResult {
+		private final boolean isTemporary;
+		private final CatalogBaseTable table;
+
+		private static TableLookupResult temporary(CatalogBaseTable table) {
+			return new TableLookupResult(true, table);
+		}
+
+		private static TableLookupResult permanent(CatalogBaseTable table) {
+			return new TableLookupResult(false, table);
+		}
+
+		private TableLookupResult(boolean isTemporary, CatalogBaseTable table) {
+			this.isTemporary = isTemporary;
+			this.table = table;
+		}
+
+		public boolean isTemporary() {
+			return isTemporary;
+		}
+
+		public CatalogBaseTable getTable() {
+			return table;
+		}
+	}
+
+	/**
 	 * Retrieves a fully qualified table. If the path is not yet fully qualified use
-	 * {@link #qualifyIdentifier(String...)} first.
+	 * {@link #qualifyIdentifier(UnresolvedIdentifier)} first.
 	 *
 	 * @param objectIdentifier full path of the table to retrieve
 	 * @return table that the path points to.
 	 */
-	public Optional<CatalogBaseTable> getTable(ObjectIdentifier objectIdentifier) {
+	public Optional<TableLookupResult> getTable(ObjectIdentifier objectIdentifier) {
 		try {
-			Catalog currentCatalog = catalogs.get(objectIdentifier.getCatalogName());
-			ObjectPath objectPath = new ObjectPath(
-				objectIdentifier.getDatabaseName(),
-				objectIdentifier.getObjectName());
-
-			if (currentCatalog != null && currentCatalog.tableExists(objectPath)) {
-				return Optional.of(currentCatalog.getTable(objectPath));
+			CatalogBaseTable temporaryTable = temporaryTables.get(objectIdentifier);
+			if (temporaryTable != null) {
+				return Optional.of(TableLookupResult.temporary(temporaryTable));
+			} else {
+				return getPermanentTable(objectIdentifier);
 			}
 		} catch (TableNotExistException ignored) {
 		}
 		return Optional.empty();
 	}
 
+	private Optional<TableLookupResult> getPermanentTable(ObjectIdentifier objectIdentifier)
+			throws TableNotExistException {
+		Catalog currentCatalog = catalogs.get(objectIdentifier.getCatalogName());
+		ObjectPath objectPath = objectIdentifier.toObjectPath();
+
+		if (currentCatalog != null && currentCatalog.tableExists(objectPath)) {
+			return Optional.of(TableLookupResult.permanent(currentCatalog.getTable(objectPath)));
+		}
+		return Optional.empty();
+	}
+
+	/**
+	 * Retrieves names of all registered catalogs.
+	 *
+	 * @return a set of names of registered catalogs
+	 */
+	public Set<String> listCatalogs() {
+		return Collections.unmodifiableSet(catalogs.keySet());
+	}
+
+	/**
+	 * Returns an array of names of all tables (tables and views, both temporary and permanent)
+	 * registered in the namespace of the current catalog and database.
+	 *
+	 * @return names of all registered tables
+	 */
+	public Set<String> listTables() {
+		return listTables(getCurrentCatalog(), getCurrentDatabase());
+	}
+
+	/**
+	 * Returns an array of names of all tables (tables and views, both temporary and permanent)
+	 * registered in the namespace of the current catalog and database.
+	 *
+	 * @return names of all registered tables
+	 */
+	public Set<String> listTables(String catalogName, String databaseName) {
+		Catalog currentCatalog = catalogs.get(getCurrentCatalog());
+
+		try {
+			return Stream.concat(
+				currentCatalog.listTables(getCurrentDatabase()).stream(),
+				listTemporaryTablesInternal(catalogName, databaseName).map(e -> e.getKey().getObjectName())
+			).collect(Collectors.toSet());
+		} catch (DatabaseNotExistException e) {
+			throw new ValidationException("Current database does not exist", e);
+		}
+	}
+
+	/**
+	 * Returns an array of names of temporary tables registered in the namespace of the current
+	 * catalog and database.
+	 *
+	 * @return names of registered temporary tables
+	 */
+	public Set<String> listTemporaryTables() {
+		return listTemporaryTablesInternal(getCurrentCatalog(), getCurrentDatabase())
+			.map(e -> e.getKey().getObjectName())
+			.collect(Collectors.toSet());
+	}
+
+	/**
+	 * Returns an array of names of temporary views registered in the namespace of the current
+	 * catalog and database.
+	 *
+	 * @return names of registered temporary views
+	 */
+	public Set<String> listTemporaryViews() {
+		return listTemporaryTablesInternal(getCurrentCatalog(), getCurrentDatabase())
+			.filter(e -> e.getValue() instanceof CatalogView)
+			.map(e -> e.getKey().getObjectName())
+			.collect(Collectors.toSet());
+	}
+
+	private Stream<Map.Entry<ObjectIdentifier, CatalogBaseTable>> listTemporaryTablesInternal(
+			String catalogName,
+			String databaseName) {
+		return temporaryTables
+			.entrySet()
+			.stream()
+			.filter(e -> {
+				ObjectIdentifier identifier = e.getKey();
+				return identifier.getCatalogName().equals(catalogName) &&
+					identifier.getDatabaseName().equals(databaseName);
+			});
+	}
+
+	/**
+	 * Lists all available schemas in the root of the catalog manager. It is not equivalent to listing all catalogs
+	 * as it includes also different catalog parts of the temporary objects.
+	 *
+	 * <b>NOTE:</b>It is primarily used for interacting with Calcite's schema.
+	 *
+	 * @return list of schemas in the root of catalog manager
+	 */
+	public Set<String> listSchemas() {
+		return Stream.concat(
+			catalogs.keySet().stream(),
+			temporaryTables.keySet().stream().map(ObjectIdentifier::getCatalogName)
+		).collect(Collectors.toSet());
+	}
+
+	/**
+	 * Lists all available schemas in the given catalog. It is not equivalent to listing databases within
+	 * the given catalog as it includes also different database parts of the temporary objects identifiers.
+	 *
+	 * <b>NOTE:</b>It is primarily used for interacting with Calcite's schema.
+	 *
+	 * @param catalogName filter for the catalog part of the schema
+	 * @return list of schemas with the given prefix
+	 */
+	public Set<String> listSchemas(String catalogName) {
+		return Stream.concat(
+			Optional.ofNullable(catalogs.get(catalogName))
+				.map(Catalog::listDatabases)
+				.orElse(Collections.emptyList())
+				.stream(),
+			temporaryTables.keySet()
+				.stream()
+				.filter(i -> i.getCatalogName().equals(catalogName))
+				.map(ObjectIdentifier::getDatabaseName)
+		).collect(Collectors.toSet());
+	}
+
+	/**
+	 * Checks if there is a catalog with given name or is there a temporary object registered within a
+	 * given catalog.
+	 *
+	 * <b>NOTE:</b>It is primarily used for interacting with Calcite's schema.
+	 *
+	 * @param catalogName filter for the catalog part of the schema
+	 * @return true if a subschema exists
+	 */
+	public boolean schemaExists(String catalogName) {
+		return getCatalog(catalogName).isPresent() ||
+			temporaryTables.keySet()
+				.stream()
+				.anyMatch(i -> i.getCatalogName().equals(catalogName));
+	}
+
+	/**
+	 * Checks if there is a database with given name in a given catalog or is there a temporary
+	 * object registered within a given catalog and database.
+	 *
+	 * <b>NOTE:</b>It is primarily used for interacting with Calcite's schema.
+	 *
+	 * @param catalogName filter for the catalog part of the schema
+	 * @param databaseName filter for the database part of the schema
+	 * @return true if a subschema exists
+	 */
+	public boolean schemaExists(String catalogName, String databaseName) {
+		return temporaryDatabaseExists(catalogName, databaseName) || permanentDatabaseExists(catalogName, databaseName);
+	}
+
+	private boolean temporaryDatabaseExists(String catalogName, String databaseName) {
+		return temporaryTables
+			.keySet()
+			.stream()
+			.anyMatch(i -> i.getCatalogName().equals(catalogName) && i.getDatabaseName().equals(databaseName));
+	}
+
+	private boolean permanentDatabaseExists(String catalogName, String databaseName) {
+		return getCatalog(catalogName)
+			.map(c -> c.databaseExists(databaseName))
+			.orElse(false);
+	}
+
 	/**
 	 * Returns the full name of the given table path, this name may be padded
-	 * with current catalog/database name based on the {@code paths} length.
+	 * with current catalog/database name based on the {@code identifier's} length.
 	 *
-	 * @param path Table path whose format can be "catalog.db.table", "db.table" or "table"
-	 * @return An array of complete table path
+	 * @param identifier an unresolved identifier
+	 * @return a fully qualified object identifier
 	 */
-	public ObjectIdentifier qualifyIdentifier(String... path) {
-		if (path == null) {
-			throw new ValidationException("Table paths can not be null!");
-		}
-		if (path.length < 1 || path.length > 3) {
-			throw new ValidationException("Table paths length must be " +
-				"between 1(inclusive) and 3(inclusive)");
-		}
-		if (Arrays.stream(path).anyMatch(StringUtils::isNullOrWhitespaceOnly)) {
-			throw new ValidationException("Table paths contain null or " +
-				"while-space-only string");
-		}
-
-		String catalogName;
-		String dbName;
-		String tableName;
-		if (path.length == 3) {
-			catalogName = path[0];
-			dbName = path[1];
-			tableName = path[2];
-		} else if (path.length == 2) {
-			catalogName = getCurrentCatalog();
-			dbName = path[0];
-			tableName = path[1];
-		} else {
-			catalogName = getCurrentCatalog();
-			dbName = getCurrentDatabase();
-			tableName = path[0];
-		}
-
-		return ObjectIdentifier.of(catalogName, dbName, tableName);
+	public ObjectIdentifier qualifyIdentifier(UnresolvedIdentifier identifier) {
+		return ObjectIdentifier.of(
+			identifier.getCatalogName().orElseGet(this::getCurrentCatalog),
+			identifier.getDatabaseName().orElseGet(this::getCurrentDatabase),
+			identifier.getObjectName());
 	}
 
 	/**
@@ -286,12 +456,68 @@ public class CatalogManager {
 	}
 
 	/**
+	 * Creates a temporary table in a given fully qualified path.
+	 *
+	 * @param table The table to put in the given path.
+	 * @param objectIdentifier The fully qualified path where to put the table.
+	 * @param replace controls what happens if a table exists in the given path,
+	 *                if true the table is replaced, an exception will be thrown otherwise
+	 */
+	public void createTemporaryTable(
+			CatalogBaseTable table,
+			ObjectIdentifier objectIdentifier,
+			boolean replace) {
+		temporaryTables.compute(objectIdentifier, (k, v) -> {
+			if (v != null && !replace) {
+				throw new ValidationException(String.format("Temporary table %s already exists", objectIdentifier));
+			} else {
+				return table;
+			}
+		});
+	}
+
+	/**
+	 * Qualifies the given {@link UnresolvedIdentifier} with current catalog & database and
+	 * removes a temporary table registered with this path if it exists.
+	 *
+	 * @param identifier potentially unresolved identifier
+	 * @return true if a table with a given identifier existed and was removed, false otherwise
+	 */
+	public boolean dropTemporaryTable(UnresolvedIdentifier identifier) {
+		return dropTemporaryTableInternal(identifier, (table) -> table instanceof CatalogTable);
+	}
+
+	/**
+	 * Qualifies the given {@link UnresolvedIdentifier} with current catalog & database and
+	 * removes a temporary view registered with this path if it exists.
+	 *
+	 * @param identifier potentially unresolved identifier
+	 * @return true if a view with a given identifier existed and was removed, false otherwise
+	 */
+	public boolean dropTemporaryView(UnresolvedIdentifier identifier) {
+		return dropTemporaryTableInternal(identifier, (table) -> table instanceof CatalogView);
+	}
+
+	private boolean dropTemporaryTableInternal(
+			UnresolvedIdentifier unresolvedIdentifier,
+			Predicate<CatalogBaseTable> filter) {
+		ObjectIdentifier objectIdentifier = qualifyIdentifier(unresolvedIdentifier);
+		CatalogBaseTable catalogBaseTable = temporaryTables.get(objectIdentifier);
+		if (filter.test(catalogBaseTable)) {
+			temporaryTables.remove(objectIdentifier);
+			return true;
+		} else {
+			return false;
+		}
+	}
+
+	/**
 	 * Alters a table in a given fully qualified path.
 	 *
 	 * @param table The table to put in the given path
 	 * @param objectIdentifier The fully qualified path where to alter the table.
 	 * @param ignoreIfNotExists If false exception will be thrown if the table or database or catalog to be altered
-	 * does not exist.
+	 *                          does not exist.
 	 */
 	public void alterTable(CatalogBaseTable table, ObjectIdentifier objectIdentifier, boolean ignoreIfNotExists) {
 		execute(
@@ -306,9 +532,14 @@ public class CatalogManager {
 	 *
 	 * @param objectIdentifier The fully qualified path of the table to drop.
 	 * @param ignoreIfNotExists If false exception will be thrown if the table or database or catalog to be altered
-	 * does not exist.
+	 *                          does not exist.
 	 */
 	public void dropTable(ObjectIdentifier objectIdentifier, boolean ignoreIfNotExists) {
+		if (temporaryTables.containsKey(objectIdentifier)) {
+			throw new ValidationException(String.format(
+				"Temporary table with identifier '%s' exists. Drop it first before removing the permanent table.",
+				objectIdentifier));
+		}
 		execute(
 			(catalog, path) -> catalog.dropTable(path, ignoreIfNotExists),
 			objectIdentifier,
